requirements.txt
streamlit 
pyyaml 
igraph
plotly